Add seperate delete expenditure view

+++ minted/views/expenditure_views.py
@@ -6,15 +6,20 @@ from .general_user_views.login_view_functions import *
 from django.contrib import messages
 
 @login_required
-def category_expenditures(request, category_name):
+def category_expenditures_view(request, category_name):
     category = Category.objects.get(user=request.user, name=category_name)
     expenditures = Expenditure.objects.filter(category=category).order_by('-date')
-    if request.method == 'POST':
-        expenditure_id = int(request.POST['id'])
-        expenditure = expenditures.get(id=expenditure_id)
-        expenditure.delete()
-        return redirect('expenditures', category_name=category_name)
     return render(request, 'expenditures/expenditures_list.html', { 'expenditures': expenditures, 'category': category })
+
+@login_required
+def delete_expenditure(request, expenditure_id):
+    if request.method == 'POST':
+        expenditure = Expenditure.objects.get(pk=expenditure_id)
+        category = expenditure.category
+        if request.user == category.user:
+            expenditure.delete()
+        return redirect('expenditures', category_name=category.name)
+    return redirect('category_list')
 
 @login_required
 def edit_expenditure(request, category_name, expenditure_id):
